T-016: PR-3: Move ai_summary_section_settings.tsx to v2 + legacy re-export
